+++ Data/SBV.hs
@@ -137,9 +137,9 @@ module Data.SBV (
   , sAnd, sOr, sAny, sAll
   -- ** Bit-vectors
   -- *** Unsigned bit-vectors
-  , SWord8, SWord16, SWord32, SWord64
+  , SWord8, SWord16, SWord32, SWord64, SWord
   -- *** Signed bit-vectors
-  , SInt8, SInt16, SInt32, SInt64
+  , SInt8, SInt16, SInt32, SInt64, SInt
   -- ** Unbounded integers
   -- $unboundedLimitations
   , SInteger
@@ -169,8 +169,8 @@ module Data.SBV (
   -- ** Single value
   -- $createSym
   , sBool, sBool_
-  , sWord8, sWord8_, sWord16, sWord16_, sWord32, sWord32_, sWord64, sWord64_
-  , sInt8,  sInt8_,  sInt16,  sInt16_,  sInt32,  sInt32_,  sInt64,  sInt64_
+  , sWord8, sWord8_, sWord16, sWord16_, sWord32, sWord32_, sWord64, sWord64_, sWordN, sWordN_
+  , sInt8,  sInt8_,  sInt16,  sInt16_,  sInt32,  sInt32_,  sInt64,  sInt64_, sIntN, sIntN_
   , sInteger, sInteger_
   , sReal, sReal_
   , sFloat, sFloat_
@@ -354,6 +354,7 @@ import Data.SBV.Core.Model      hiding (assertWithPenalty, minimize, maximize,
                                         sWord8, sWord8_, sWord8s, sWord16, sWord16_, sWord16s,
                                         sWord32, sWord32_, sWord32s, sWord64, sWord64_, sWord64s,
                                         sMaybe, sMaybe_, sMaybes, sEither, sEither_, sEithers, sSet, sSet_, sSets)
+import Data.SBV.Core.Sized
 import Data.SBV.Core.Floating
 import Data.SBV.Core.Splittable
 import Data.SBV.Core.Symbolic   (MonadSymbolic(..), SymbolicT)