+++ CHANGES.md
@@ -5,6 +5,12 @@
 
 ### Version 8.3.5, Not yet released
 
+  * SBV now supports arbitrary-size bit-vectors, i.e.,
+    `SWord 17`, `SInt 9` etc. These should work like any
+    other bit-vector, using the `DataKinds` feature of
+    GHC. Thanks to Ben Blaxill for the idea and the initial
+    implementation.
+
   * Ramped up travis-appveyor build infrastructure.
 
 ### Version 8.3, 2019-06-08